test: Animal/Person controllers

+++ src/main/java/com/diginamic/demo/repository/SpeciesRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.diginamic.demo.entity.Species;
 
 @Repository
-public interface SpeciesRepository extends JpaRepository<Species, Long> {
+public interface SpeciesRepository extends JpaRepository<Species, Integer> {
 	public Species findFirstByCommonName(final String commonName);
 	public List<Species> findByLatinNameIgnoreCaseContains(final String latinName);
 }
